Refactor MapClient import path and remove unused components for cleaner structure
